requirements.txt
bcrypt==5.0.0
cffi==2.0.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
fonttools==4.61.1
invoke==2.2.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
paramiko==4.0.0
pillow==12.1.0
pyadi-iio==0.0.20
pycparser==3.0
pylibiio==0.25
PyNaCl==1.6.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
pyyaml>=6.0
scipy>=1.10.0
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.17.0
six==1.17.0
